refactor: move everything related to network into folder
